hydra refactor
This is a larger commit which begins a refactor to using hydra for script config.
Conceptually, config arguments can now be separated into `estimate` arguments which need to be provided for any subtraction method, and `phorc` arguments which are specific to using phorc.

Example of this is in conf/run_phorc_config.yaml.
Later, we can also add arguments which are specific to CAVIaR

Todo:
- refactor other scripts to also use hydra
- (optional) separate config scripts to pull from separate files for subtraction, caviar, etc.